backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent-backend"
version = "0.1.0"
description = "AI Agent Backend with Docker Support"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi==0.103.1",
    "uvicorn==0.23.2",
    "pydantic==2.3.0",
    "openai==1.2.0",
    "httpx==0.27.2",
    "python-dotenv==1.0.0",
    "python-multipart==0.0.6",
    "docker==6.1.3",
]

[project.optional-dependencies]
dev = [
    "black==23.1.0",
    "isort==5.12.0",
    "mypy==1.1.1",
    "pytest==7.3.1",
    "pytest-cov==4.1.0",
]

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false